+++ src/app/domains/statistics/analytics.model.ts
@@ -1,0 +1,38 @@
+import { Phenophase, Species } from '@shared/models/masterdata.model';
+import { Observation } from '@shared/models/observation.model';
+
+export type AltitudeGroup = 'alt1' | 'alt2' | 'alt3' | 'alt4' | 'alt5';
+export type AltitudeFilterGroup = 'all' | AltitudeGroup;
+export type AnalyticsType = 'species' | 'altitude';
+
+export const allPhenophases = { id: 'all', de: 'Alle' } as Phenophase;
+export const allSpecies = { id: 'all', de: 'Alle' } as Species;
+export const allYear = 'all';
+
+export class AnalyticsValue {
+  phenophase: string;
+  max: Date;
+  median: Date;
+  min: Date;
+  quantile_25: Date;
+  quantile_75: Date;
+}
+
+export class Analytics {
+  source: string;
+  species: string;
+  type: AnalyticsType;
+  altitude_grp: AltitudeGroup | null;
+  year: number;
+  values: AnalyticsValue[];
+}
+
+export interface ObservationData {
+  species: string;
+  groupedByPhenophase: GroupedByPhenophaseGroup[];
+}
+
+export interface GroupedByPhenophaseGroup {
+  phenophaseGroup: string;
+  observations: Observation[];
+}
