+++ src/app/domains/statistics/statistics.service.ts
@@ -3,12 +3,11 @@ import { Firestore, where } from '@angular/fire/firestore';
 import { FormControl, FormGroup } from '@angular/forms';
 import { BaseResourceService } from '@core/services/base-resource.service';
 import { FirestoreDebugService } from '@core/services/firestore-debug.service';
-import { Phenophase } from '@shared/models/masterdata.model';
 import { SourceFilterType } from '@shared/models/source-type.model';
 import { Observable } from 'rxjs';
 import { map, tap } from 'rxjs/operators';
-import { Statistics } from '../../shared/models/statistics';
-import { AltitudeFilterGroup, AnalyticsType } from './statistics.model';
+import { AltitudeFilterGroup, AnalyticsType } from './analytics.model';
+import { Statistics } from './statistics.model';
 
 @Injectable({ providedIn: 'root' })
 export class StatisticsService extends BaseResourceService<Statistics> {
